requirements.txt
flask
flask-cors
python-dotenv
google-genai
google-adk
uvicorn
asgiref
requests
docling
easyocr
pypdf
scrapy==2.11.0
playwright==1.40.0
html2text==2020.1.16
beautifulsoup4==4.12.0
lxml==4.9.3
chromadb
sentence-transformers
